修改：Long类型

+++ websocket-item/src/main/java/com/example/websocketitem/controller/CommentController.java
@@ -24,15 +24,15 @@ public class CommentController {
         return commentService.listCommentPage(pageNum,pageSize);
     }
     @GetMapping("/levelOne")
-    public Result<List<Comment>> selectByTrendsIdAndCommentLevel(Integer articleId, Integer commentLevel){
+    public Result<List<Comment>> selectByTrendsIdAndCommentLevel(Long articleId, Integer commentLevel){
         return commentService.selectByArticleIdIdAndCommentLevel(articleId, commentLevel);
     }
     @GetMapping("/levelTwo")
-    public Result<List<Comment>> selectByParentCommentIdAndTrendsIdAndCommentLevel(Integer parentCommentId, Integer articleId, Integer commentLevel){
+    public Result<List<Comment>> selectByParentCommentIdAndTrendsIdAndCommentLevel(Long parentCommentId, Long articleId, Integer commentLevel){
         return commentService.selectByParentCommentIdAndArticleIdIdAndCommentLevel(parentCommentId, articleId, commentLevel);
     }
     @PutMapping("/praise")
-    public Result<Comment> updatePraiseNumByCommentId(Integer commentId, Integer praiseNum){
+    public Result<Comment> updatePraiseNumByCommentId(Long commentId, Integer praiseNum){
         return commentService.updatePraiseNumByCommentId(commentId, praiseNum);
     }
     @PostMapping
@@ -44,12 +44,12 @@ public class CommentController {
         return commentService.deleteByPrimaryKey(id);
     }
     @GetMapping("/userId")
-    public Result<PageInfo<Comment>> getCommentByUserId(@RequestParam Integer userId, @RequestParam(defaultValue = "1") int pageNum,
+    public Result<PageInfo<Comment>> getCommentByUserId(@RequestParam Long userId, @RequestParam(defaultValue = "1") int pageNum,
                                            @RequestParam(defaultValue = "10") int pageSize){
         return commentService.selectAllByUserId(userId,pageNum,pageSize);
     }
     @GetMapping("/tree")
-    public Result<List<Tree<Integer>>> listCommentALlTree(){
+    public Result<List<Tree<Long>>> listCommentALlTree(){
         return commentService.listCommentAll();
     }
     @GetMapping("/oneWeek")
